refactor(cli): extract bench output printers to a sibling module
